requirements.txt
asgiref==3.6.0
Django==4.2
django-restframework==0.0.1
django-seed==0.3.1
djangorestframework==3.14.0
Faker==18.3.2
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
toposort==1.10
tzdata==2023.3
